feat: extract slot components separately
