+++ package/src/components/MessageInput/MessageComposerTrailingView.tsx
@@ -1,0 +1,3 @@
+export const MessageComposerTrailingView = () => {
+  return null;
+};
