+++ package/src/components/MessageInput/utils/messageComposerSelectors.ts
@@ -1,0 +1,9 @@
+import { MessageComposerState, TextComposerState } from 'stream-chat';
+
+export const textComposerStateSelector = (state: TextComposerState) => ({
+  command: state.command,
+});
+
+export const messageComposerStateStoreSelector = (state: MessageComposerState) => ({
+  quotedMessage: state.quotedMessage,
+});
